Compress JSON in URL's
The new sippy produces some hideous URL's. There's not a lot we can do
about it, since it's stateless (no remembering short URL's) and we have
some complex filtering needs. However, we can make them a little less
unsightly. This moves to using JSONCrush, an algorithm for reducing the
size of JSON by more than half, while still also being URL safe and
somewhat readable.
