+++ sippy-ng/src/releases/ReleaseOverview.js
@@ -10,10 +10,10 @@ import SimpleBreadcrumbs from '../components/SimpleBreadcrumbs'
 import SummaryCard from '../components/SummaryCard'
 import { BOOKMARKS, INFRASTRUCTURE_THRESHOLDS, INSTALL_THRESHOLDS, UPGRADE_THRESHOLDS } from '../constants'
 import VariantCards from '../jobs/VariantCards'
-import { queryForBookmark } from '../helpers'
 import JobTable from '../jobs/JobTable'
 import TestTable from '../tests/TestTable'
 import PropTypes from 'prop-types'
+import { queryForBookmark } from '../lib/urls'
 
 export const TOOLTIP = 'Top level release indicators showing product health'
 export const REGRESSED_TOOLTIP = 'Shows the most regressed items this week vs. last week, for those with more than 10 runs'
